server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb==0.5.5",
    "openai==1.43.0",
    "orjson==3.10.7",
    "pandas==2.2.2",
    "python-dotenv==1.0.1",
    "tqdm==4.66.4",
]
